fixes done-1

+++ database/migrations/2024_03_30_182916_create_trax_booking_pieces_table.php
@@ -16,7 +16,7 @@ class CreateTraxBookingPiecesTable extends Migration
         Schema::create('trax_booking_pieces', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('booking_id');
-            $table->integer('piece_cn_number');
+            $table->bigInteger('piece_cn_number');
 //            $table->integer('from_pieces')->nullable();
 //            $table->integer('to_pieces')->nullable();
             $table->integer('scan_rider_id')->nullable();
